ml-service/requirements.txt
tensorflow
grpcio
grpcio-tools
protobuf
numpy
